% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrophoto_bar.R
\name{spectrophoto_bar}
\alias{spectrophoto_bar}
\title{Spectophotography bar graph}
\usage{
spectrophoto_bar(data, x, y)
}
\arguments{
\item{data}{the desired dataset}

\item{x}{the x-axis variable}

\item{y}{the y-axis variable}
}
\value{
bar graph
}
\description{
The spectrophoto_bar function easily gives an overview of your spectrophotography measurements, per group.
}
\details{
Note: for the example, no actual spectophotography data was used.
It's just an example of what the output should look like.
}
\examples{
spectrophoto_bar(iris, iris$Species, iris$Sepal.Length)
}
