% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrophoto_boxplot.R
\name{spectrophoto_boxplot}
\alias{spectrophoto_boxplot}
\title{Spectrophotography boxplot}
\usage{
spectrophoto_boxplot(data, group, y)
}
\arguments{
\item{data}{the desired dataset}

\item{group}{the desired group}

\item{y}{the extinction data}
}
\value{
boxplot
}
\description{
The spectrophoto_boxplot function makes it easier to quickly generate a boxplot on your spectrophotography data.
}
\details{
Note: for the example, no actual spectophotography data was used.
It's just an example of what the output should look like.
}
\examples{
spectrophoto_boxplot(iris, iris$Species, iris$Sepal.Length)
}
