% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_table.R
\name{url_table}
\alias{url_table}
\title{Adding an url column to your data}
\usage{
url_table(dataset, column, urls)
}
\arguments{
\item{dataset}{The desired dataset}

\item{column}{The desired column (dataset$column)}

\item{urls}{The urls you want to insert}
}
\value{
dataset with url column
}
\description{
url_table easily adds a column with the desired urls, per cell!
}
\examples{
url_table(iris, iris$Species, urls = c(rep(c("https://en.wikipedia.org/wiki/Iris_setosa"), times = 50), rep(c("https://en.wikipedia.org/wiki/Iris_versicolor"), times = 50), rep(c("https://en.wikipedia.org/wiki/Iris_virginica"), times = 50)))
}
