% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_graph.R
\name{blast_graph}
\alias{blast_graph}
\title{Make a graph from BLAST+ output}
\usage{
blast_graph(path, reads)
}
\arguments{
\item{path}{Path to a BLAST+ output file}

\item{reads}{Amount of reads you want to filter by}
}
\value{
column graph
}
\description{
This function automatically generates a graph from BLAST+ input
This way you'll see more easily what the composition is like.
}
